safari-600.7-branch/LayoutTests/fast/regions/cssom/webkit-named-flow-overset-expected.txt

Test for https://bugs.webkit.org/show_bug.cgi?id=87491 [CSS Regions]Rename NamedFlow::overflow to NamedFlow::overset. For more details see http://www.w3.org/TR/2012/WD-css3-regions-20120503/

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS namedFlow.overset is true
PASS namedFlow.overset is false
PASS namedFlow.overset is true
PASS namedFlow.overset is false
PASS namedFlow.overset is false
PASS namedFlow.overset is false
PASS namedFlow.overset is false
PASS namedFlow.overset is false
PASS namedFlow.overset is true
PASS successfullyParsed is true

TEST COMPLETE
